Part A is working and is debuged.

+++ include/vslam_lib.h
@@ -5,4 +5,4 @@ void send_pose_information(sw::redis::Redis* redis, const slamcore::Map2DInterfa
               const slamcore::PoseInterface<slamcore::camera_clock>& pose);
 bool check_map_data(sw::redis::Redis* redis);
 void feed_encoder_data(std::string msg, std::shared_ptr<slamcore::MobileRobotSubsystemInterface> robot_feed, slamcore::IDT* sample_counter);
-int modulo(int a, int b);
+double modulo(double a, double b);
